requirements-dev.txt
# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-metadata==3.0.0
pytest-sugar==0.9.7
pytest-timeout==2.2.0
pytest-rerunfailures==14.0
pytest-randomly==3.15.0
pytest-benchmark==4.0.0
pytest-docker==2.0.1
pytest-env==1.1.3
pytest-forked==1.6.0
pytest-instafail==0.5.0
pytest-repeat==0.9.3

# Linting & Formatting
pylint==3.0.2
black==23.9.1
isort==5.12.0
mypy==1.6.1

# Type stubs
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10
types-pyyaml==6.0.12.12
types-pytz==2023.3.1.1

# Additional development dependencies
transformers>=4.30.0
torch>=2.0.1
nltk>=3.8.1
pyyaml>=6.0.2
requests>=2.32.3
msal>=1.22.0
python-docx>=1.1.0
